pyproject.toml
[project]
name = "murmur"
version = "0.2.0"
description = "Record system audio from meetings (Zoom, Google Meet, etc.) via PipeWire + FFmpeg"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "click>=8.3.1",
    "rich>=14.3.3",
]

[project.optional-dependencies]
transcribe = ["faster-whisper>=1.0.0"]
summarize = []
diarize = ["pyannote-audio>=3.0", "torch"]
all = ["murmur[transcribe]", "murmur[diarize]"]
dev = ["pytest>=8.0", "ruff>=0.11", "pre-commit>=4.0"]

[project.scripts]
murmur = "murmur.cli:cli"

[project.entry-points."murmur.plugins"]
transcribe = "murmur.plugins.transcribe:register"
summarize = "murmur.plugins.summarize:register"
diarize = "murmur.plugins.diarize:register"
tui = "murmur.plugins.tui:register"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py314"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
